Split out ijruAverage to avoid pulling in FS into speed bundle
