+++ lib/models/competition-events/ijru.freestyle@3.0.0.test.ts
@@ -1,6 +1,7 @@
 import assert from 'node:assert'
 import test from 'node:test'
 import * as mod from './ijru.freestyle@3.0.0.js'
+import { ijruAverage } from './ijru.common.js'
 import { type JudgeResult, type EntryMeta, type JudgeMeta } from '../types.js'
 import { RSRWrongJudgeTypeError } from '../../errors.js'
 
@@ -26,31 +27,31 @@ void test('ijru.freestyle@3.0.0', async t => {
 
   await t.test('ijruAverage', async t => {
     await t.test('Should return single number', () => {
-      assert.strictEqual(mod.ijruAverage([1]), 1)
+      assert.strictEqual(ijruAverage([1]), 1)
     })
 
     await t.test('Should average two numbers', () => {
-      assert.strictEqual(mod.ijruAverage([1, 3]), 2)
+      assert.strictEqual(ijruAverage([1, 3]), 2)
     })
 
     await t.test('Should average the closest two of three numbers, when the lower two are closest', () => {
-      assert.strictEqual(mod.ijruAverage([1, 10, 3]), 2)
+      assert.strictEqual(ijruAverage([1, 10, 3]), 2)
     })
 
     await t.test('Should average the closest two of three numbers, when the higher two are closest', () => {
-      assert.strictEqual(mod.ijruAverage([1, 10, 8]), 9)
+      assert.strictEqual(ijruAverage([1, 10, 8]), 9)
     })
 
     await t.test('Should average the highest two of three numbers, when the numbers are equidistant', () => {
-      assert.strictEqual(mod.ijruAverage([1, 1 + 3, 1 + 3 + 3]), 5.5)
+      assert.strictEqual(ijruAverage([1, 1 + 3, 1 + 3 + 3]), 5.5)
     })
 
     await t.test('Should average all except highest and lowest for four numbers', () => {
-      assert.strictEqual(mod.ijruAverage([119, 114, 111, 118]), 116)
+      assert.strictEqual(ijruAverage([119, 114, 111, 118]), 116)
     })
 
     await t.test('Should average all except highest and lowest for five numbers', () => {
-      assert.strictEqual(mod.ijruAverage([119, 114, 131, 111, 118]), 117)
+      assert.strictEqual(ijruAverage([119, 114, 131, 111, 118]), 117)
     })
   })
