+++ lib/models/competition-events/ijru.freestyle@3.0.0.ts
@@ -1,6 +1,7 @@
 import { RSRWrongJudgeTypeError } from '../../errors.js'
 import { filterTally, formatFactor, matchMeta, roundTo, roundToCurry, simpleCalculateTallyFactory } from '../../helpers.js'
 import type { CompetitionEventModel, JudgeFieldDefinition, JudgeTypeGetter, TableDefinition } from '../types.js'
+import { ijruAverage } from './ijru.common.js'
 
 type Option = 'noMusicality' | 'discipline' | 'interactions'
 
@@ -20,28 +21,6 @@ export function L (l: number): number {
   return roundTo(0.1 * Math.pow(1.5, l), 2)
 }
 
-export const ijruAverage = (scores: number[]): number => {
-  // sort ascending
-  scores.sort(function (a, b) {
-    return a - b
-  })
-
-  if (scores.length >= 4) {
-    scores.pop()
-    scores.shift()
-
-    const score = scores.reduce((a, b) => a + b)
-    return score / scores.length
-  } else if (scores.length === 3) {
-    const closest = scores[1] - scores[0] < scores[2] - scores[1] ? scores[1] + scores[0] : scores[2] + scores[1]
-    return closest / 2
-  } else if (scores.length === 2) {
-    const score = scores.reduce((a, b) => a + b)
-    return score / scores.length
-  } else {
-    return scores[0]
-  }
-}
 
 // ======
 // JUDGES
