+++ lib/models/competition-events/ijru.speed@1.0.0.ts
@@ -1,7 +1,7 @@
 import { RSRWrongJudgeTypeError } from '../../errors.js'
 import { filterTally, matchMeta, roundTo, simpleCalculateTallyFactory } from '../../helpers.js'
 import type { CompetitionEventModel, JudgeTypeGetter, TableDefinition } from '../types.js'
-import { ijruAverage } from './ijru.freestyle@3.0.0.js'
+import { ijruAverage } from './ijru.common.js'
 
 type Option = 'falseSwitches'
