+++ package.json
@@ -20,11 +20,18 @@
         "default": "./dist/esm/lib/index.js"
       }
     },
-    "./models/competition-events/*": {
-        "types": "./dist/esm/lib/models/competition-events/*.d.ts",
-        "default": "./dist/esm/lib/models/competition-events/*.js"
+    "./models/*": {
+      "require": {
+        "types": "./dist/cjs/lib/models/*.d.ts",
+        "default": "./dist/cjs/lib/models/*.js"
+      },
+      "import": {
+        "types": "./dist/esm/lib/models/*.d.ts",
+        "default": "./dist/esm/lib/models/*.js"
       }
+    }
   },
+  "sideEffects": false,
   "files": [
     "dist/"
   ],
